+++ piccolo/apps/playground/commands/run.py
@@ -233,7 +233,7 @@ def populate():
             RecordingStudio.facilities: {
                 "restaurant": True,
                 "mixing_desk": True,
-                "instruments": {"electric_guitar": 10, "drum_kit": 2},
+                "instruments": {"electric_guitars": 10, "drum_kits": 2},
             },
         }
     )
@@ -245,7 +245,7 @@ def populate():
             RecordingStudio.facilities: {
                 "restaurant": False,
                 "mixing_desk": True,
-                "instruments": {"electric_guitar": 6, "drum_kit": 3},
+                "instruments": {"electric_guitars": 6, "drum_kits": 3},
             },
         },
     )
