+++ piccolo/query/functions/__init__.py
@@ -1,13 +1,11 @@
 from .aggregate import Avg, Count, Max, Min, Sum
 from .datetime import Day, Extract, Hour, Month, Second, Strftime, Year
-from .json import Arrow
 from .math import Abs, Ceil, Floor, Round
 from .string import Concat, Length, Lower, Ltrim, Reverse, Rtrim, Upper
 from .type_conversion import Cast
 
 __all__ = (
     "Abs",
-    "Arrow",
     "Avg",
     "Cast",
     "Ceil",
